Move resolution constants from front to core::util::dims
