+++ benches/e2e.rs
@@ -9,7 +9,7 @@ use retrofire_core::{
         translate, viewport,
     },
     render::{Context, Frag, Model, debug::dir_to_rgb, render, shader},
-    util::{Buf2, Dims, pnm},
+    util::{Buf2, Dims, dims, pnm},
 };
 use retrofire_geom::solids::{Build, Sphere};
 
@@ -32,7 +32,7 @@ fn triangle(b: Bencher, n: u32) {
         |frag: Frag<Color3f<_>>, _: &_| frag.var.to_color4(),
     );
 
-    let dims @ Dims(w, h) = Dims(640, 480);
+    let dims @ Dims(w, h) = dims::VGA_640_480;
     let modelview = translate((0.0, 0.0, 2.0)).to();
     let project = perspective(1.0, w as f32 / h as f32, 0.1..1000.0);
     let viewport = viewport(pt2(0, h)..pt2(w, 0));
@@ -76,7 +76,7 @@ fn sphere(b: Bencher, res: u32) {
         |frag: Frag<Color4f>, _: &_| frag.var.to_color4(),
     );
 
-    let dims @ Dims(w, h) = Dims(640, 480);
+    let dims @ Dims(w, h) = dims::VGA_640_480;
     let modelview = translate((0.0, 0.0, 2.0)).to();
     let project = perspective(1.0, w as f32 / h as f32, 0.1..1000.0);
     let viewport = viewport(pt2(0, h)..pt2(w, 0));
